Feature/tga

+++ src/libsail/sail_advanced.c
@@ -25,6 +25,7 @@
 
 #include "config.h"
 
+#include <stddef.h>
 #include <stdlib.h>
 
 #include "sail-common.h"
@@ -127,7 +128,7 @@ sail_status_t sail_read_next_frame(void *state, struct sail_image **image) {
     }
 
     /* Allocate pixels. */
-    const unsigned pixels_size = image_local->height * image_local->bytes_per_line;
+    const size_t pixels_size = (size_t)image_local->height * image_local->bytes_per_line;
     SAIL_TRY_OR_CLEANUP(sail_malloc(pixels_size, &image_local->pixels),
                         /* cleanup */ sail_destroy_image(image_local));
